requirements.txt
discord
python-dotenv
openai
faiss-cpu 
tiktoken 
langchain
langchain-community
langchain_openai